Patch:Linux 3.4.11-3.4.105
Linux Upstream patches upto 3.4.105. Credits to original Maintainer Li Zefan and its former maintainer Greg Kroah-Hartman.

Highlights:
- Upstream patches
- Just reverted "futex: Unlock hb->lock in futex_wait_requeue_pi() error path"
- Ommitted some patches as they were already cherry-picked or manually added
- Compilation fixes and resynched device configuration

+++ include/linux/msg.h
@@ -59,13 +59,14 @@ struct msginfo {
 #ifdef __KERNEL__
 #include <linux/list.h>
 
+/* one msg_msg structure for each message */
 struct msg_msg {
-	struct list_head m_list; 
-	long  m_type;          
-	int m_ts;           
+	struct list_head m_list;
+	long m_type;
+	size_t m_ts;		/* message text size */         
 	struct msg_msgseg* next;
 	void *security;
-	
+	/* the actual message follows immediately */
 };
 
 struct msg_queue {
